+++ Makefile
@@ -4885,7 +4885,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/completion_queue.h \
     include/grpcpp/completion_queue_impl.h \
     include/grpcpp/create_channel.h \
-    include/grpcpp/create_channel_impl.h \
     include/grpcpp/create_channel_posix.h \
     include/grpcpp/ext/health_check_service_server_builder_option.h \
     include/grpcpp/generic/async_generic_service.h \
@@ -4969,7 +4968,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/auth_context.h \
     include/grpcpp/security/auth_metadata_processor.h \
     include/grpcpp/security/credentials.h \
-    include/grpcpp/security/credentials_impl.h \
     include/grpcpp/security/server_credentials.h \
     include/grpcpp/security/server_credentials_impl.h \
     include/grpcpp/security/tls_credentials_options.h \
@@ -5577,7 +5575,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/completion_queue.h \
     include/grpcpp/completion_queue_impl.h \
     include/grpcpp/create_channel.h \
-    include/grpcpp/create_channel_impl.h \
     include/grpcpp/create_channel_posix.h \
     include/grpcpp/ext/health_check_service_server_builder_option.h \
     include/grpcpp/generic/async_generic_service.h \
@@ -5661,7 +5658,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/auth_context.h \
     include/grpcpp/security/auth_metadata_processor.h \
     include/grpcpp/security/credentials.h \
-    include/grpcpp/security/credentials_impl.h \
     include/grpcpp/security/server_credentials.h \
     include/grpcpp/security/server_credentials_impl.h \
     include/grpcpp/security/tls_credentials_options.h \
