+++ include/grpcpp/security/server_credentials_impl.h
@@ -81,7 +81,7 @@ std::shared_ptr<ServerCredentials> LocalServerCredentials(
 
 /// Builds TLS ServerCredentials given TLS options.
 std::shared_ptr<ServerCredentials> TlsServerCredentials(
-    const TlsCredentialsOptions& options);
+    const grpc::experimental::TlsCredentialsOptions& options);
 
 }  // namespace experimental
 }  // namespace grpc_impl
