+++ src/cpp/client/cronet_credentials.cc
@@ -55,10 +55,9 @@ class CronetChannelCredentialsImpl final : public ChannelCredentials {
   }
   void* engine_;
 };
-}  // namespace grpc
-namespace grpc_impl {
+
 std::shared_ptr<ChannelCredentials> CronetChannelCredentials(void* engine) {
   return std::shared_ptr<ChannelCredentials>(
       new grpc::CronetChannelCredentialsImpl(engine));
 }
-}  // namespace grpc_impl
+}  // namespace grpc
