+++ src/cpp/common/tls_credentials_options.cc
@@ -23,7 +23,7 @@
 #include "src/core/lib/security/credentials/tls/grpc_tls_credentials_options.h"
 #include "src/cpp/common/tls_credentials_options_util.h"
 
-namespace grpc_impl {
+namespace grpc {
 namespace experimental {
 
 /** TLS key materials config API implementation **/
@@ -340,4 +340,4 @@ TlsCredentialsOptions::TlsCredentialsOptions(
 TlsCredentialsOptions::~TlsCredentialsOptions() {}
 
 }  // namespace experimental
-}  // namespace grpc_impl
+}  // namespace grpc
