+++ src/cpp/server/secure_server_credentials.cc
@@ -94,7 +94,7 @@ void AuthMetadataProcessorAyncWrapper::InvokeProcessor(
 
 }  // namespace grpc
 
-namespace grpc_impl {
+namespace grpc {
 
 int SecureServerCredentials::AddPortToServer(const std::string& addr,
                                              grpc_server* server) {
@@ -149,11 +149,11 @@ std::shared_ptr<ServerCredentials> LocalServerCredentials(
 }
 
 std::shared_ptr<ServerCredentials> TlsServerCredentials(
-    const TlsCredentialsOptions& options) {
+    const grpc::experimental::TlsCredentialsOptions& options) {
   grpc::GrpcLibraryCodegen init;
   return std::shared_ptr<ServerCredentials>(new SecureServerCredentials(
       grpc_tls_server_credentials_create(options.c_credentials_options())));
 }
 
 }  // namespace experimental
-}  // namespace grpc_impl
+}  // namespace grpc
